Fix case where unknown function is nested
Because we didn't replace the pipes, unknown commands in the
parameters had their {{ }} replaced by placeholders, but the
pipes remained, which meant that when the upper function
was evaluated, the code thought that the inner unknown-func
parameters are the ones for the outer function.

Fixed by making the pipes also be replaced by placeholder in
the case where we have unknown function that we are skipping
and then restore placeholders.

Note that in the initial run (before we process) to replace
single brackets with placeholders, we do NOT replace pipes too
because that case is only for single-brackets; if we replace
all pipes there, we'd break all functions.