[BREAKING CHANGE] Add the ability to work with async functions
This is a breaking change, since it would mean that the process
returns a promise rather than the direct string. The rest of the
operation remains the same.

This "small" (though breaking) functionality can allow for a much
more robust set of functions to be inserted, like functions that
fetch remote data. An example will be added to docs.